feat: 실수 계산 지원 추가
- Numbers 클래스를 double 타입으로 변경하여 정수와 실수 모두 지원
- Calculator, CalculatorService, Controller, OutputView를 double 타입으로 변경
- 정수인 경우 소수점 없이, 실수인 경우 소수점 포함하여 출력
- 정수 범위 초과 시 예외 처리 유지 (Integer.MAX_VALUE)

+++ src/main/java/calculator/controller/CalculatorController.java
@@ -28,7 +28,7 @@ public class CalculatorController {
     public void run() {
         try {
             String input = InputView.readInput();
-            int result = calculatorService.calculate(input);
+            double result = calculatorService.calculate(input);
             OutputView.printResult(result);
         } catch (IllegalArgumentException e) {
             ErrorView.printError(e.getMessage());
